Fix db commit problem
While working on the branch I discovered that an earlier change from the optimser work resulted in no data being persisted and the transaction being thrown due to an commit error. I ended up having to change the properties.

The following [link](https://dzone.com/articles/how-to-delay-connection-acquisition-until-its-real) describes the solution.