Fix DocumentDescription BSM support
The BSM work was not finalised previously so BSM support for DocumentDescription was only partially complete. While fixing this I noticed that we were using a setReferenceBSM approach on a list. This should be avoided especially on updates as the persistence session will see the list as a new set of objects not the already known set of objects. Therefore the set approach was taken away and the Bsm List and other instances were set as final to reflect the approach we want to use.