Drop unofficial convertFil (konverterFil) end point
Noark 5 Tjenestegrensesnitt do not specify any need for manually triggering conversion, and
conversion can be done externally using the official endpoints, so I believe there is no need
for this endpoint.

If it should be possible to request a new convers, or ask Nikita to try again to do the
conversion, I believe it would be better to use PUT, PATCH or some other operation on the
reference file (referanseFil) end point.