Make first letter lowercase
In general all variables in the entity classes start with a lowercase. Author is both an entity and a class. So when creating a foreign key join, referenceAuthor will be used so "Author" requires a capital letter. When it comes to the author variable, we require a lowercase "author". This change has not been required before because we have not had a situation like Author/author.  The entire string is not being made lowercase as there are plenty of variables following camelCase.