Fix Repository inheritance
These should be paging repository so that when we bring in pagination these entities will support this. Probably originally I expected these entities to not require pagination. That is why they didn't support it. This change comes now as when I discovered it I know not fixing it will surprise me later.