Refactor find class object from a string
The ability to get an instance of a class from its string name is also useful when determining which HateoasNoarkObject to instantiate. In fact the previous code had a potential type identification if the [0] element in the array was a subclass and not the parent class. So this also fixes a previous unknown bug in the code. What's happening here is to move the generation of a Class list of domain model entities to startup rather than being called everytime there is a OData query to be solved.