Major refactoring to support pagination
Once basic pagination was solved and need to be rolled out there was a need to refactor the code so that all endpoints could make use of pagination. There have been a lot of tight coupling in the codebase that makes refactoring in nikita difficult. This commit does the following:
  * roll out pagination across codebase (minus Metadata)
  * standardises the controller/service code
  * reduce the number of warnings in the codebase
  * makes future refactoring easier as coupling is reduced

 Note: Metadata entities are still not paginated. That comes in a later commit.
 The coupling made it difficult to find single good points to commit changes.  Therefore it was easier to just do a big commit.
 Casefile expansion is currently not wokring as it is being pushed through the same code as CaseFile and failing on a test.