Fix problem deserialising screeningmetadata
There were two problems. The first is that we were looking for an object called _skjermingmetadata_ rather than the correct value (according to the spec) that is _skjermingMetadata_. there are a few such inconsistencies in nikita eand the API standard that can be difficult to spot. Next the null check on if meant that we never removed an empty screeningmetata object and the deserialising failed.