Remove SAST
While it is interesting to see the results in the pipeline, we simply do not have capacity to follow this up at the moment. As such we are generating a lot of information unnecessary. We are using coverity for static analysis and when we catch up with everything, we can start bring the SAST back into the pipeline.