Include angular-img-http-src in html
common_functionality.js has the declaration of  angular.module('nikita' ... This means that any page that uses  common_functionality.js will require the angular-http to be loaded. This is being brought in as when we want to load an image from nikita we will require code that can add the required http tokens to the request to get the image. This library provides such support, but the way nikita-gui is built means that it will need to be included in all html files.