Update composite key approach
When the hibernate library was updated, error started occurring during startup. I believe this to be related to the earlier way composite keys were generated. Rather than trying to fix that approach I decided to reimplement it using the approach described [here](https://vladmihalcea.com/the-best-way-to-map-a-composite-primary-key-with-jpa-and-hibernate/). This resulted in some minor updates of missing constants and I added constant values where applicable. Also includes updated column name used in tests