Add indexing to relevant domain model attributes
I have tried to be consistent with what is fulltextfield / keywordfield / genericfield. This is a first attempt and I expect we may make changes in the future when we get to know ES a little better. No weighting is added now either.

Note: An index is created per object type. This is default behavior from hibernate-search.

Note: Metadata values and codename etc are not being indexed in this round. The reason is that I am not sure if it is worthwhile, and I want to keep the approach as simple as possible.