Update the CLOB bridge
While debugging, we came across various problems that actually were not problems but the index had not been updated. Searching for a solution to a non-existent problem led me to:

https://stackoverflow.com/questions/22874722/clob-to-string-bufferedreader-vs-getsubstring

I like the approach and reasoning behind the description so we have adapted the solution detailed in the SO post above.