Remove unused @ActiveProfiles annotation
I don't think the "test" profile is defined anywhere, so there's no need to use this in the ActiveProfiles annotation. Using this annotation was also overriding the profile specified at the command line (spring.profiles.active), which we need to be able to enable the "ci" profile when using the Gitlab runner.

But just in case it IS necessary we will add "test" as an additional profile on the command line (see .gitlab-ci.yml).